kim2026/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kim2026"
version = "0.1.0"
description = "Reproducible near-IR turbulence beam-cleanup D2NN simulator"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "torch>=2.2",
  "matplotlib>=3.8",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[project.scripts]
kim2026-generate-pairs = "kim2026.cli.generate_pairs:main"
kim2026-train = "kim2026.cli.train_beam_cleanup:main"
kim2026-evaluate = "kim2026.cli.evaluate_beam_cleanup:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
